+++ changelog.md
@@ -10,6 +10,8 @@ Codion Change Log
 - SwingMcpServer.takeApplicationScreenshot() reimplemented, bypassing the robot, window no longer has to be in front.
 - SwingMcpServer, window listing tool added and window screenshot functionality improved, refactoring.
 - SwingMcpPlugin.startMcpServer() removed, mcpServer() returning State added.
+### is.codion.swing.common.model
+- FilterTableModel.Builder.filterModelFactory() renamed filters().
 ### is.codion.swing.common.ui
 - FilterTable.Builder.sortingEnabled() renamed sortable().
 ### is.codion.framework.db.core